Complete myhyperpowers-11e: build shared Pi task runner and migrate hyperpowers_subagent
- Add shared runner with fresh/fork context support
- Migrate subagent execution to runner-backed adapter
- Add runner tests for context and abort behavior

Part of epic: myhyperpowers-sf3 - Add Hyperpowers-native Pi orchestration core and skill metadata
